ft-agent-backend/requirements.txt
# Universal Agent Backend - Requirements

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL driver

# AI & LLM
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.20
openai==1.12.0

# Document Processing
pypdf==4.0.1

# Text Processing
langchain-text-splitters==0.0.1

# Vector Store
qdrant-client==1.7.3

# Embeddings
sentence-transformers==2.3.1

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Auth
PyJWT==2.8.0

# File Upload
python-multipart==0.0.9

# Token Counting
tiktoken==0.7.0
